push CoroutineContext interface down to BasePageFragment
